Update SortingVisualizer.py
 --- SORTING VISUALIZER ---

Introduction
To visualize and understand how famous sorting algorithms work. A sorting algorithm is an algorithm that organizes elements of a sequence in a certain order. This project provides two standpoints to look at algorithms, one on visual display and frontend design, the other is more analytical at providing proficient logic to develop the algorithm.  Algorithms differ on methods and therefore efficiency. With each sorting method the user can see the steps by which the algorithm analyses and sorts the data. Providing insight into their speed and efficiency. 

The visual aspect and design of this project was not paramount, apologies to any artistic designers. I used Tkinter to develop the small interactive GUI, incorporating a minimalistic 2D design. This GUI was development to make simplistic, interactive GUI where a user can alter data and sorting method. Thereby visualizes how each sorting algorithm works on datasets. 

Sorting Algorithms 
A sorting algorithm is an algorithm made up of a series of instructions that takes an array as input, performs specified operations on the array, sometimes called a list, and outputs a sorted array. This is usually performed on alphabetical or numerical datasets. For this project I used a numerical dataset. Sorting algorithms provide a straightforward way in introducing other key computer science and software engineering topics like Big-O notation, divide-and-conquer methods, and data structures such as binary trees.

There are some factors in choosing when choosing the right algorithm for a particular problem. These are based on primarily on running time, space complexity and the format of the input list. For example, Quick sort is a very fast algorithm but tricky to implement, whereas Bubble sort is easy implement but also very slow. The choice may then come down to the dataset and its size. For a small dataset it may be best to quickly implement the Bubble sort algorithm that can handle this size. The opposite would be true for larger datasets, where the time taken to implement the Quick sort algorithm would prove beneficial.

Bubble Sort - is a comparison​-based algorithm that compares each pair of elements in an array and swaps them if they are out of order until the entire array is sorted. For each element in the list, the algorithm compares every pair of elements.

Merge Sort - is a comparison-based algorithm that focuses on how to merge together two pre-sorted arrays such that the resulting array is also sorted. It first splits the array into single elements and two partitions. It compares each pair in both partitions, and then compares partitions. 

Quick Sort - is a comparison-based algorithm that uses divide-and-conquer to sort an array. It works by selecting a 'pivot' element from the array and partitioning the other elements into two sub-arrays, according to whether they are less than or greater than the pivot. The sub-arrays are then sorted recursively.

Tkinter - provides a variety of common GUI elements which we can use to build our interface – such as buttons, menus and various kinds of entry fields and display areas such as canvas and a foundation this is 'Frame'.  These are tools are libraries with specific features that can adjusted to develop a particular design. It involves common features that are found in front-end web development such as grid, padding, row and column. I also utilised other unique features such as a Scale for algorithm speed and data size, and Combobox for algorithm menu.

The tkinter library is called as 'root' and this is where configuration begins. In order to generate data, draw the data and select and algorithm to work on the data, three helper functions were developed.

Generate - pulls the inputs from the user and uses these numbers to determine the data size and data range. This is the volume of data in the data set and its range. 

Start_algo - uses an if/else statement to decide on what choice has been made by the user. This function then calls the draw_data function to draw and represent this data on the canvas (GUI). 

Draw_data - then uses the determined data parameters and algorithm displaying it on the canvas for visualisation. First the canvas is set, the data set then normalised to fit inside on the canvas (fit to scale). This is performed with a for loop that offsets the data versus the canvas, depending on data input. The smaller the data set the larger the bars and vice versa. The canvas the updates each time an even occurs through a tkinter too (update_idletasks()).

Different colours are used in this help function and with each sorting method to show moving data, pivot points and overall sorting visualisation. Finally root.mainloop() executes the application.