Update TicTacToe+AI.py
--- AI & Tic Tac Toe ---

INTRODUCTION

Tic tac toe is a two player game typically played on a 3x3 square. This game size can increase to 4x4, 5x5 and so on. Of the two players, one plays as 'X', the other as 'O'. The X player also starts and winner is determined when one player completes a full line horizontally, vertically or diagonally on the game board. If neither player can make a complete line the game ends in a draw.

The aim of this project was to implement a robust version of this game with an unbeatable computer player, denoted as AI (artificial intelligence). This AI player uses a specific algorithm to assess all possible moves on the game board and determine the best outcome to prevent the opponent from winner and therefore promotes its chances to win. This algorithm is called the MiniMax algorithm. 


MINIMAX
This algorithm is applied in many two player games such as chess, checkers and go. These games are known as zero-sum games, because in a mathematical representation: one player wins (+1) and other player loses (-1) or both of anyone not to win (0). The algorithm recursively searches the current game state to find the best move leading the 'Max' player to win. The recursive search stores best possible outcomes in a 'best_simulation' variable, which updates when a new current best move is found. A best move is the shortest route to victory. The algorithm must work in the constraints of all available moves that can be played. It searches from null to infinity to determine this best possible outcome. 

The algorithm searches all possible moves as a game tree. The current position is the root and possible moves are branches that can be scored as a +1 for a win or -1 for a loss. Alpha-beta pruning is also implemented to shorten this search or lookup. Alpha-beta pruning is a search algorithm that seeks to decrease the number of nodes (branch points) that are evaluated by the minimax algorithm in its search tree.