Update TicTacToe+AI.py
--- AI & Tic Tac Toe ---


INTRODUCTION

Tic tac toe is a two player game typically played on a 3x3 square. This game size can increase to 4x4, 5x5 and so on. Of the two players, one plays as 'X', the other as 'O'. The X player also starts and winner is determined when one player completes a full line horizontally, vertically or diagonally on the game board. If neither player can make a complete line the game ends in a draw.

The aim of this project was to implement a robust version of this game with an unbeatable computer player, denoted as AI (artificial intelligence). This AI player uses a specific algorithm to assess all possible moves on the game board and determine the best outcome to maximize it's chances at winning by minimizing it's opponents chances. This algorithm is called the MiniMax algorithm. 


MINIMAX

Minimax is a decision rule used in artificial intelligence, decision theory, game theory, statistics, and philosophy for minimizing the possible loss for a worst case scenario. Minimax is a kind of backtracking algorithm that assumes your opponent also plays optimally This algorithm is applied in many two player games such as chess, checkers and go. In Minimax the two players are called Maximizer and Minimizer. The Maximizer tries to get the highest score possible while the Minimizer tries to do the opposite and get the lowest score possible. These games are known as zero-sum games, because in a mathematical representation: one player wins (+1, Maximizer) and other player loses (-1, Minimizer) or if neither cases are true (0).

Since this is a backtracking based algorithm, it tries all possible moves, then backtracks and makes a decision. The algorithm recursively searches the current game state to find the best move leading the 'Max' player to win and 'Min' player to lose. It searches from null to infinity on all possible outcomes, or null to the relevant recursion dept (board size).  The recursive search stores the best possible outcomes in a 'best_simulation' variable, which updates when a new current best move is found. A best move is the shortest route to victory or minimum loss. The algorithm must work in the constraints of all available moves that can be played on the current board state.

The algorithm searches all possible moves as a game tree. The current position is the root. The 'Max' searches left and right of the root. Then the 'Min' searches each of the left and right nodes, always choosing the minimum value. The 'Max' then chooses the larger of these two values. In this way it is backtracking to find the shortest route to victory/best move. Alpha-beta pruning is implemented to shorten this search or lookup. Alpha-beta pruning is an optimization technique that seeks to decrease the number of nodes (branch points) that are evaluated in the search tree. Hereby decreasing the entire algorithms time complexity. It cuts off branches in the game tree which need not be searched because there already exists a better move available. Alpha is the best value that the Maximizer currently can guarantee at that level or above. Beta is the best value that the Minimizer currently can guarantee at that level or above.