Update MovieRecommender.py
-- MOVIE RECOMMENDER --

Introduction to Machine learning, Pandas, Sklearn and Numpy. Followed and programmed simple instructions to develop an application that read data (movie file) that are stored using key value pairing. Each movie (data) has keywords in its pairing. These keywords are used to pair with user input, checking for similarity and thereby returning recommendations.  The application uses Cosine Similarity to weigh and compare key words/features in movies.

Machine Learning
Machine learning is a subset of artificial intelligence, it is the science of getting computers to act without being explicitly programmed, and is mostly just statistics. It is used to find patterns in data that you can then make predictions on.

Pandas
Pandas is a software library written for the Python programming language for data manipulation and analysis. In particular, it offers data structures and operations for manipulating numerical tables and time series. It is built on the Numpy package and its key data structure is called the DataFrame. DataFrames allow you to store and manipulate tabular data in rows of observations and columns of variables.


Numpy
Sklearn


Cosine Similarity