Update SortingVisualizer.py
-- SORTING VISUALIZER -- 

The purpose of this project is to develop an interactive GUI to visualize sorting algorithms working on a dataset. A sorting algorithm is an algorithm made up of a series of instructions that takes an array as input, performs specified operations on the array, sometimes called a list, and outputs a sorted array. The array can be sorted in alphabetical or numerical order. I used a numerical dataset in this project. The primary reason I chose to do this project is due to sorting algorithms providing a straightforward way to introduce other key computer science topics like Big-O notation, divide-and-conquer methods, and data structures such as binary trees, and heaps. The essence of efficient programming and design. 

All sorting algorithms share the goal of outputting a sorted list, but the way that each algorithm goes about this task can vary. It is important to know how fast it runs and in how much space it operates—in other words, its time complexity and space complexity. Also, implementing a more efficient algorithm can often become more complex and take longer, as opposed to an easier, less efficient algorithm. It is important to take into account the size of the dataset (how much space is needed to run the algorithm, space complexity) and running time (how many operations the algorithm must perform before completing, time complexity). 

For example, Quick sort is a very fast algorithm but can be pretty tricky to implement; Bubble sort is a slow algorithm but is very easy to implement. To sort small sets of data, Bubble sort may be a better option since it can be implemented quickly, but for larger datasets, the speedup/efficiency from Quick sort might be worth the trouble implementing the algorithm.


Bubble Sort - O(n 2)

Bubble sort is a comparison​-based algorithm that compares each pair of elements in an array and swaps them if they are out of order until the entire array is sorted. For each element in the list, the algorithm compares every pair of elements.


Merge Sort - O(n log n)

Merge sort is a comparison-based algorithm that focuses on how to merge together two pre-sorted arrays such that the resulting array is also sorted. It first recursively separates the entire array in single elements, and partitions the array to left and right parts. It then compares and swaps elements in each partition. Finally, it swaps the remaining elements in the partitions until the array is sorted.


Quick sort - O(n log n), although can be unstable

Quick sort is a comparison sort, meaning that it can sort items of any type for which a "less-than" relation is defined. It is a divide-and-conquer technique that works by selecting a 'pivot' element from the array and partitioning the other elements into two sub-arrays, according to whether they are less than or greater than the pivot. The sub-arrays are then sorted recursively. 


Tkinter - 

The GUI for this project is implemented using Tkinter. Tkinter is the standard GUI library for Python. It provides a fast and easy way to create GUI applications, as the front-end development of this project was not the focus. The design is a 2D minimalistic with intuitive interactive features. It provides simply dropdown boxes and start/generate buttons. As well as scales for data size and speed. 

As the algorithm works on the dataset, the GUI changes the color of the relevant data elements to illustrate, the swapping elements, pivot points, partitions and completed data set. These features are implemented using 3 helper functions as follows;


Generate - 

This functions takes in the user input for data size, min and max values. Then calls the draw data function to display this on the GUI formatting all elements in the color 'yellow'.

Draw data - 

Draw data uses the canvas methods from tkinter to clear, set and create the canvas in which the data will be displayed. I had to normalize the data so that it fits to scale within the canvas. This means that the data size, min and max entries, will effectively alter the data bars drawn onto the canvas. The larger the dataset the smaller the drawn bars and vice versa. Another tkinter method, update_idletasks(), is used to constantly update the function once it has completed an iteration. 

Start algorithm - 
 
The start_algo function uses an if/else statement or clause to take the user input from the dropdown box to get the chosen algorithm. Then calling the draw-data function to perform using this algorithm, changing the generated data from 'yellow' to the live/working data 'purple'.